chore(use-common): move types to qeta-common package
