+++ plugins/qeta-common/src/types.ts
@@ -1,0 +1,21 @@
+export interface StatisticResponse {
+  ranking: Statistic[];
+  loggedUser?: Statistic;
+}
+
+export interface Statistic {
+  author?: string;
+  total?: number;
+  position?: string;
+}
+
+export interface StatisticsOptions {
+  limit?: number;
+  period?: string;
+  loggedUser?: string;
+}
+
+export interface StatisticsRequestParameters {
+  author?: string;
+  options?: StatisticsOptions;
+}
