+++ plugins/qeta/src/api/QetaClient.ts
@@ -11,12 +11,14 @@ import {
   QuestionResponseBody,
   QuestionsResponse,
   QuestionsResponseBody,
-  StatisticResponse,
-  StatisticsRequestParameters,
   TagResponse,
 } from './types';
 import omitBy from 'lodash/omitBy';
 import isEmpty from 'lodash/isEmpty';
+import {
+  StatisticResponse,
+  StatisticsRequestParameters,
+} from '@drodil/backstage-plugin-qeta-common';
 
 export const qetaApiRef = createApiRef<QetaApi>({
   id: 'plugin.qeta.service',
