+++ plugins/qeta/src/api/types.ts
@@ -108,25 +108,3 @@ export interface AttachmentResponse {
 }
 
 export type AttachmentResponseBody = AttachmentResponse | ErrorResponse;
-
-export interface StatisticResponse {
-  ranking: Statistic[];
-  loggedUser?: Statistic;
-}
-
-export interface Statistic {
-  author?: string;
-  total?: number;
-  position?: string;
-}
-
-export interface StatisticsOptions {
-  limit?: number;
-  period?: string;
-  loggedUser?: string;
-}
-
-export interface StatisticsRequestParameters {
-  author?: string;
-  options?: StatisticsOptions;
-}
